chore(cluster-init): refactor sanitizeprowjob code
