+++ cmd/cluster-init/cmd/onboard/sanitizeprowjob/update_test.go
@@ -1,4 +1,4 @@
-package onboard
+package sanitizeprowjob
 
 import (
 	"testing"
